Variable agent setup
Moving toward variable number of agents

+++ Planning/AStar_easy.h
@@ -22,7 +22,7 @@ using namespace boost;
 using namespace std;
 using namespace easymath;
 
-typedef float cost;
+typedef double cost;
 
 // euclidean distance heuristic
 template <class Graph, class CostType, class LocMap>
@@ -30,7 +30,7 @@ class distance_heuristic : public astar_heuristic<Graph, CostType>
 {
 public:
 	typedef typename graph_traits<Graph>::vertex_descriptor Vertex;
-	distance_heuristic(LocMap l, Vertex goal, int xdim, int ydim): 
+	distance_heuristic(LocMap l, Vertex goal, double xdim, double ydim): 
 		m_location(l), 
 		m_goal(goal),
 		XDIM(xdim),
@@ -49,7 +49,7 @@ public:
 private:
 	LocMap m_location;
 	Vertex m_goal;
-	int XDIM, YDIM;
+	double XDIM, YDIM;
 };
 
 
@@ -84,7 +84,11 @@ public:
 class AStar_easy
 {
 public:
-	typedef adjacency_list<listS, vecS, undirectedS, no_property,
+	typedef adjacency_list
+		<listS, // edge container
+		vecS,	// vertex container
+		undirectedS,	// graph type is undirected -- SHOULD CHANGE THIS
+		no_property,
 		property<edge_weight_t, cost> > mygraph_t;
 	typedef property_map<mygraph_t, edge_weight_t>::type WeightMap;
 	typedef mygraph_t::vertex_descriptor vertex;
@@ -228,7 +232,7 @@ public:
 	static void ind2sub(int cols, int ind, int &r, int &c){
 		//0-indexed: ind-1 always called
 		r = (ind-1)%cols;
-		c = floor((ind-1)/cols);
+		c = int(floor((ind-1)/cols));
 	}
 	int XDIM, YDIM; // dimensions of the map
 	set<XY> nodes; // set of nodes already found
